Moved two peak player count queries to TPSQueries class
